refactor: split wallet-ui.js into 15 focused submodules
Break down the monolithic wallet-ui.js (5600 lines, 179 functions) into
focused modules under src/renderer/lib/wallet/. The coordinator stays as
a thin ~320-line file that initializes submodules and re-exports the
public API — no changes needed in consumer files (index.js, dapp-provider.js).
